Added Webdriver exception handler and --sessionperspec option
Webdriver exceptions that are not handled by the spec itself are
now properly caught by Webdriver's control flow and reported as
a failed expectation.

Note it is still doable to create code that throws an exception
that will make the whole process crash, but there is basically
nothing that we can do about it. For instance, this will crash:

describe('A crash', function () {
  it('is a crash', function (done) {
    setTimeout(function () {
      throw new Error('Crash!');
    }, 0);
  });
});

The --sessionperspec option makes the test runner create one
browser session per spec file. This makes things run extremely
slow, so use only if needed!