Added bits on "black-box testing"
There are many projects around that run unit tests written with
Jasmine in PhantomJS. That is not the purpose of Angela.

Angela is useful to test Web applications from an end user
perspective: the application is considered to be a black box and
the tests emulate the actions of an end user to test the reactions
of that black box.