Detailed error message when an async spec test uses sync version
It's easy to forget the "done" parameter when writing an async test.
Unfortunately, this makes Jasmine crash for the time being and thus
makes the whole process crash.

Angela now correctly catches that situation, cleans things up as
needed and reports the error with a message that warns the user
about the likely cause of the exception.